Pin the forward to the calculator dtype
MACECalculator scoped default_dtype around batch construction only, so the
model forward ran under whatever dtype the process had set. Extensions build
tensors during the forward without an explicit dtype, les among them for its
3x3 identities, and those followed the process default instead of the model.
The mismatch survived the forward through type promotion and only failed in
the backward, as "linalg.lu_solve: Expected LU and B to have the same dtype".

Scope the forward too, and give the two identities in extensions.py an
explicit dtype so the model is also correct when called directly.

The regression test previously depended on an earlier test in the same xdist
worker leaving the default at float64, which made it an intermittent failure.
It now sets the mismatch itself.

+++ mace/modules/extensions.py
@@ -503,7 +503,11 @@ class MACELES(ScaleShiftMACE):
                         hasattr(self, "use_anisotropic_polarizability")
                         and self.use_anisotropic_polarizability
                     ):
-                        eye = torch.eye(3, device=node_alphas.device)
+                        eye = torch.eye(
+                            3,
+                            device=node_alphas.device,
+                            dtype=node_alphas.dtype,
+                        )
                         node_alphas = node_alphas.unsqueeze(-1).unsqueeze(
                             -1
                         ) * eye.unsqueeze(0)
@@ -543,7 +547,7 @@ class MACELES(ScaleShiftMACE):
             )
             # Make quads traceless:
             traces = les_quad.diagonal(dim1=-1, dim2=-2).sum(dim=1)
-            eye = torch.eye(3, device=les_quad.device)
+            eye = torch.eye(3, device=les_quad.device, dtype=les_quad.dtype)
             les_quad = les_quad - eye[None, :, :] * traces[:, None, None] / 3
         else:
             les_quad = None
